+++ planner/core/BUILD.bazel
@@ -170,7 +170,6 @@ go_test(
         "expression_rewriter_test.go",
         "expression_test.go",
         "find_best_task_test.go",
-        "flat_plan_test.go",
         "fragment_test.go",
         "indexmerge_intersection_test.go",
         "indexmerge_path_test.go",
@@ -231,6 +230,7 @@ go_test(
         "//parser/mysql",
         "//parser/terror",
         "//planner",
+        "//planner/core/internal",
         "//planner/property",
         "//planner/util",
         "//session",
@@ -240,8 +240,6 @@ go_test(
         "//sessiontxn",
         "//statistics",
         "//statistics/handle",
-        "//store/mockstore",
-        "//store/mockstore/unistore",
         "//table",
         "//testkit",
         "//testkit/ddlhelper",
@@ -270,12 +268,10 @@ go_test(
         "@com_github_golang_snappy//:snappy",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
-        "@com_github_pingcap_kvproto//pkg/metapb",
         "@com_github_pingcap_tipb//go-tipb",
         "@com_github_prometheus_client_golang//prometheus",
         "@com_github_prometheus_client_model//go",
         "@com_github_stretchr_testify//require",
-        "@com_github_tikv_client_go_v2//testutils",
         "@org_golang_x_exp//slices",
         "@org_uber_go_goleak//:goleak",
     ],
