+++ planner/core/casetest/BUILD.bazel
@@ -8,7 +8,11 @@ go_test(
         "cbo_test.go",
         "enforce_mpp_test.go",
         "expression_rewriter_test.go",
+        "flat_plan_test.go",
+        "integration_partition_test.go",
+        "integration_test.go",
         "main_test.go",
+        "physical_plan_test.go",
         "plan_test.go",
         "point_get_plan_test.go",
         "predicate_simplification_test.go",
@@ -29,10 +33,12 @@ go_test(
         "//expression",
         "//expression/aggregation",
         "//infoschema",
+        "//kv",
         "//parser",
         "//parser/ast",
         "//parser/model",
         "//parser/mysql",
+        "//parser/terror",
         "//planner",
         "//planner/core",
         "//planner/core/internal",
@@ -44,6 +50,7 @@ go_test(
         "//statistics",
         "//statistics/handle",
         "//testkit",
+        "//testkit/external",
         "//testkit/testdata",
         "//testkit/testmain",
         "//testkit/testsetup",
